requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.4
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distro==1.5.0
dj-inmemorystorage==2.1.0
Django==3.1.6
django-adminlte3==0.1.6
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-concurrency==2.2
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.0
django-cruds-adminlte==0.0.16
django-dynamic-formsets==0.0.8
django-filter==2.4.0
django-glrm==1.1.3
django-image-cropping==1.5.0
django-import-export==2.5.0
django-tablib==3.2
django-widgets-improved==1.5.0
djangoajax==3.2
docker==4.4.3
docker-compose==1.28.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
factory-boy==3.2.0
Faker==6.0.0
flake8==3.8.4
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.9
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
tenacity==6.3.1
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.3
uWSGI==2.0.19.1
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
